update runtime error test
