--- test/error-runtime/error-runtime-exploring/error-runtime-exploring.js
@@ -1,3 +1,0 @@
-import { triggerError } from "./trigger-error.js"
-
-triggerError()
