--- test/error-runtime/error-runtime-late-launch-browser/error-runtime-late-launch-browser.js
@@ -1,5 +1,0 @@
-import { triggerError } from "./trigger-error.js"
-
-setTimeout(() => {
-  triggerError()
-}, 1000)
