--- test/error-runtime/error-runtime-late-launch-node/error-runtime-late-launch-node.js
@@ -1,5 +1,0 @@
-import { triggerError } from "./trigger-error.js"
-
-setTimeout(() => {
-  triggerError()
-}, 1000)
