--- test/error-runtime/error-runtime-late-launch-node/trigger-error.js
@@ -1,3 +1,0 @@
-export const triggerError = () => {
-  throw new Error("error")
-}
