--- test/error-runtime/error-runtime-launch-browser/error-runtime-launch-browser.html
@@ -1,11 +1,0 @@
-<html>
-  <head>
-    <title>Title</title>
-    <meta charset="utf-8" />
-    <link rel="icon" href="data:," />
-  </head>
-
-  <body>
-    <script type="module" src="./error-runtime-launch-browser.js"></script>
-  </body>
-</html>
