--- test/error-runtime/error-runtime-launch-browser/error-runtime-launch-browser.js
@@ -1,3 +1,0 @@
-import { triggerError } from "./trigger-error.js"
-
-triggerError()
