--- test/error-runtime/error-runtime-launch-node/error-runtime-launch-node.js
@@ -1,3 +1,0 @@
-import { triggerError } from "./trigger-error.js"
-
-triggerError()
