--- test/error-runtime/error-runtime-launch-node/trigger-error.js
@@ -1,3 +1,0 @@
-export const triggerError = () => {
-  throw new Error("SPECIAL_STRING_UNLIKELY_TO_COLLIDE")
-}
