--- test/error-runtime/error-runtime-testing-node/file.js
@@ -1,1 +1,0 @@
-export const ask = () => 40
