--- test/error-runtime/error-runtime-testing-node/file.spec.html
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <title>Title</title>
-    <meta charset="utf-8" />
-    <link rel="icon" href="data:," />
-  </head>
-
-  <body>
-    <script type="module" src="./file.spec.js"></script>
-  </body>
-</html>
