--- test/error-runtime/error-runtime-testing-node/file.spec.js
@@ -1,7 +1,0 @@
-import { ask } from "./file.js"
-
-const actual = ask()
-const expected = 42
-if (actual !== expected) {
-  throw new Error(`ask() should return ${expected}, got ${actual}`)
-}
