+++ test/error_runtime/error_runtime.html
@@ -1,4 +1,3 @@
-<!DOCTYPE html>
 <html>
   <head>
     <title>Title</title>
@@ -7,6 +6,6 @@
   </head>
 
   <body>
-    <script type="module" src="./error-runtime-exploring.js"></script>
+    <script type="module" src="./error_runtime.js"></script>
   </body>
 </html>
