+++ test/error_runtime/error_runtime.js
@@ -1,0 +1,3 @@
+import { triggerError } from "./trigger_error.js"
+
+triggerError()
