+++ test/error_runtime/error_runtime_after_timeout/error_runtime_after_timeout.html
@@ -7,6 +7,6 @@
   </head>
 
   <body>
-    <script type="module" src="./error-runtime-late-launch-browser.js"></script>
+    <script type="module" src="./error_runtime_after_timeout.js"></script>
   </body>
 </html>
