+++ test/error_runtime/error_runtime_after_timeout/error_runtime_after_timeout.js
@@ -1,0 +1,5 @@
+import { triggerError } from "./trigger_error.js"
+
+setTimeout(() => {
+  triggerError()
+}, 1000)
