+++ test/error_runtime/error_runtime_launch_node.test.js
@@ -1,9 +1,5 @@
 import { assert } from "@jsenv/assert"
-import {
-  resolveDirectoryUrl,
-  urlToRelativeUrl,
-  urlToBasename,
-} from "@jsenv/filesystem"
+import { resolveUrl, urlToRelativeUrl } from "@jsenv/filesystem"
 
 import { launchNode } from "@jsenv/core"
 import { jsenvCoreDirectoryUrl } from "@jsenv/core/src/internal/jsenvCoreDirectoryUrl.js"
@@ -15,14 +11,13 @@ import {
   LAUNCH_TEST_PARAMS,
 } from "@jsenv/core/test/TEST_PARAMS_LAUNCH_NODE.js"
 
-const testDirectoryUrl = resolveDirectoryUrl("./", import.meta.url)
+const testDirectoryUrl = resolveUrl("./", import.meta.url)
 const testDirectoryRelativeUrl = urlToRelativeUrl(
   testDirectoryUrl,
   jsenvCoreDirectoryUrl,
 )
-const testDirectoryname = urlToBasename(testDirectoryRelativeUrl)
 const jsenvDirectoryRelativeUrl = `${testDirectoryRelativeUrl}.jsenv/`
-const filename = `${testDirectoryname}.js`
+const filename = `error_runtime.js`
 const fileRelativeUrl = `${testDirectoryRelativeUrl}${filename}`
 const { origin: compileServerOrigin, outDirectoryRelativeUrl } =
   await startCompileServer({
@@ -30,8 +25,7 @@ const { origin: compileServerOrigin, outDirectoryRelativeUrl } =
     jsenvDirectoryRelativeUrl,
   })
 
-let errorCallbackArgValue
-const actual = await launchAndExecute({
+const result = await launchAndExecute({
   ...LAUNCH_AND_EXECUTE_TEST_PARAMS,
   launchAndExecuteLogLevel: "off",
   launch: (options) =>
@@ -44,21 +38,19 @@ const actual = await launchAndExecute({
   executeParams: {
     fileRelativeUrl,
   },
-  runtimeErrorCallback: (value) => {
-    errorCallbackArgValue = value
-  },
+  captureConsole: true,
 })
+
+const actual = {
+  status: result.status,
+  errorMessage: result.error.message,
+  consoleCallsContainsString: result.consoleCalls.some(({ text }) =>
+    text.includes("SPECIAL_STRING_UNLIKELY_TO_COLLIDE"),
+  ),
+}
 const expected = {
-  status: "completed",
-  namespace: {},
+  status: "errored",
+  errorMessage: "SPECIAL_STRING_UNLIKELY_TO_COLLIDE",
+  consoleCallsContainsString: false,
 }
 assert({ actual, expected })
-
-process.on("beforeExit", () => {
-  const actual = errorCallbackArgValue
-  const expected = {
-    error: Object.assign(new Error("child exited with 1"), { exitCode: 1 }),
-    timing: "after-execution",
-  }
-  assert({ actual, expected })
-})
